[SVLS-5262] try putting the pointer calculations in the trace_handlers instead of botocore
