+++ ddtrace/_trace/trace_handlers.py
@@ -10,7 +10,9 @@ from typing import Optional
 import wrapt
 
 from ddtrace._trace.span import Span
+from ddtrace._trace.utils import SpanPointerDescription
 from ddtrace._trace.utils import extract_DD_context_from_messages
+from ddtrace._trace.utils import extract_span_pointers_from_successful_botocore_response
 from ddtrace._trace.utils import set_botocore_patched_api_call_span_tags as set_patched_api_call_span_tags
 from ddtrace._trace.utils import set_botocore_response_metadata_tags
 from ddtrace.constants import _ANALYTICS_SAMPLE_RATE_KEY
@@ -620,7 +622,17 @@ def _on_botocore_patched_api_call_exception(ctx, response, exception_type, is_er
 
 
 def _on_botocore_patched_api_call_success(ctx, response):
-    set_botocore_response_metadata_tags(ctx.get_item(ctx.get_item("call_key")), response)
+    span = ctx.get_item(ctx.get_item("call_key"))
+
+    set_botocore_response_metadata_tags(span, response)
+
+    for span_pointer_description in extract_span_pointers_from_successful_botocore_response(
+        endpoint_name=ctx.get_item("endpoint_name"),
+        operation_name=ctx.get_item("operation"),
+        request_parameters=ctx.get_item("params"),
+        response=response,
+    ):
+        _set_span_pointer(span, span_pointer_description)
 
 
 def _on_botocore_trace_context_injection_prepared(
@@ -774,8 +786,8 @@ def _on_test_visibility_is_enabled() -> bool:
     return CIVisibility.enabled
 
 
-def _set_span_pointer(ctx, span_pointer_description) -> None:
-    ctx.get_item(ctx.get_item("call_key"))._add_span_pointer(
+def _set_span_pointer(span: Span, span_pointer_description: SpanPointerDescription) -> None:
+    span._add_span_pointer(
         pointer_kind=span_pointer_description.pointer_kind,
         pointer_direction=span_pointer_description.pointer_direction,
         pointer_hash=span_pointer_description.pointer_hash,
@@ -839,8 +851,6 @@ def listen():
     core.on("test_visibility.enable", _on_test_visibility_enable)
     core.on("test_visibility.disable", _on_test_visibility_disable)
     core.on("test_visibility.is_enabled", _on_test_visibility_is_enabled, "is_enabled")
-
-    core.on("datadog.span_pointer", _set_span_pointer)
 
     for context_name in (
         "flask.call",
